+++ src/main/kotlin/compiler/Preprocessor.kt
@@ -12,7 +12,21 @@ object Preprocessor {
     }
 
     fun preprocessText (text:String, filePath: String): String {
-        val bylines = text.trim().lines().map { it.substringBefore("//") }.toTypedArray()
+        val splitedCode = text.split("\"").toMutableList()
+
+        for (i in 0 until splitedCode.size) {
+            if (i % 2 == 0) continue
+            splitedCode[i] = splitedCode[i].replace("\\n", "\\\\n")
+            splitedCode[i] = splitedCode[i].replace("\\t", "\\t")
+            splitedCode[i] = splitedCode[i].replace(" ", "#\$#CZECHTINAMEZERA\$#\$")
+            splitedCode[i] = splitedCode[i].replace("{", "#\$#CZECHTINAOPEMN\$#\$")
+            splitedCode[i] = splitedCode[i].replace("}", "#\$#CZECHTINACLOSE\$#\$")
+        }
+        val code = splitedCode.joinToString("\"").trim()
+
+
+
+        val bylines = code.trim().lines().map { it.substringBefore("//") }.toTypedArray()
 
         var blocklevel = 0
         for (i in bylines.indices) {
@@ -38,6 +52,6 @@ object Preprocessor {
                 bylines[i] = bylines[i].substringBeforeLast("\\")
         }
 
-        return bylines.joinToString("\n")
+        return bylines.joinToString("\n").replace("#\$#CZECHTINAOPEMN\$#\$", "{").replace("#\$#CZECHTINACLOSE\$#\$", "}")
     }
 }